Add files via upload
You are given an Integer array in 'A' of size N, sorted in non-decreasing order. You are also given an integer 'target'. Your task is to write a Function to search for 'target' in array 'A'. If it exists written it's index in Zero based indexing using Linear Search in C++.